refactored the code, removed previous implementations of the performance metrics and added a dummy method for getting all of them
