+++ data-extraction-scripts/training_runner.py
@@ -1,8 +1,9 @@
 #!/usr/bin/env python3
 import subprocess
 from csv_table_creator import write_csv_row, get_csv_columns
-from data_extraction_script import parse_mlagents_output, calculate_general_data
-
+from data_extraction_script import parse_mlagents_output
+from general_data_metrics import calculate_general_data
+from performance_data_metrics import calculate_performance_data
 
 def run_training(config_file_path, run_id, chosen_game, learning_algorithm, csv_output_path):
     """
@@ -70,9 +71,11 @@ def run_training(config_file_path, run_id, chosen_game, learning_algorithm, csv_
             # If we have a step and at least one metric, and haven't written this step yet, write to CSV
             if current_step is not None and current_step not in written_steps:
                 if current_mean_reward is not None or current_std_reward is not None or current_policy_loss is not None:
-                    # Calculate general data (called on each line with metrics)
+
+                    # Calculate general, performance and resource usage data (called on each line with metrics)
                     general_data = calculate_general_data(chosen_game, learning_algorithm, run_id, current_step)
-                    
+                    performance_data = calculate_performance_data()
+
                     # Prepare row data
                     row_data = {
                         "step": current_step,
